+++ packages/upload/src/upload-list.tsx
@@ -26,7 +26,7 @@
 
 import { computed, defineComponent, h, toRefs, TransitionGroup } from 'vue';
 
-import { useLocale } from '@bkui-vue/config-provider';
+import { useLocale, usePrefix } from '@bkui-vue/config-provider';
 import { ArchiveFill, AudioFill, Del, Done, ImageFill, RightTurnLine, TextFill, VideoFill } from '@bkui-vue/icon';
 import Progress from '@bkui-vue/progress';
 import { classes } from '@bkui-vue/shared';
@@ -45,9 +45,11 @@ export default defineComponent({
   emits: ['remove', 'retry'],
   setup(props, { slots, emit }) {
     const t = useLocale('upload');
+    const { resolveClassName } = usePrefix();
+
     const { theme, disabled, multiple } = toRefs(props);
 
-    const classBlock = `${CLASS_PREFIX}-list`;
+    const classBlock = `${resolveClassName(CLASS_PREFIX)}-list`;
 
     const isPhotowall = computed<boolean>(() => theme.value === EThemes.PICTURE);
     const isSinglePhoto = computed<boolean>(() => isPhotowall.value && !multiple.value);