+++ packages/upload/src/upload.tsx
@@ -26,6 +26,7 @@
 
 import { computed, defineComponent, onBeforeUnmount, shallowRef } from 'vue';
 
+import { usePrefix } from '@bkui-vue/config-provider';
 import { classes } from '@bkui-vue/shared';
 
 import uploadProps from './props';
@@ -50,6 +51,8 @@ export default defineComponent({
   props: uploadProps,
   emits: ['exceed', 'progress', 'success', 'error', 'delete', 'done'],
   setup(props, { slots, emit, expose }) {
+    const { resolveClassName } = usePrefix();
+
     const requests = shallowRef<Record<string, XMLHttpRequest | Promise<unknown>>>({});
 
     const isPhotowall = computed<boolean>(() => props.theme === EThemes.PICTURE);
@@ -57,10 +60,10 @@ export default defineComponent({
 
     const classNames = computed(() =>
       classes({
-        [CLASS_PREFIX]: true,
-        [`${CLASS_PREFIX}--${props.theme}`]: true,
-        [`${CLASS_PREFIX}--disabled`]: props.disabled,
-        [`${CLASS_PREFIX}--single-picture`]: isSinglePhoto.value,
+        [`${resolveClassName(CLASS_PREFIX)}`]: true,
+        [`${resolveClassName(CLASS_PREFIX)}--${props.theme}`]: true,
+        [`${resolveClassName(CLASS_PREFIX)}--disabled`]: props.disabled,
+        [`${resolveClassName(CLASS_PREFIX)}--single-picture`]: isSinglePhoto.value,
         [props.extCls]: props.extCls ?? false,
       }),
     );
@@ -247,7 +250,7 @@ export default defineComponent({
             onChange={handleFiles}
           />
         )}
-        {slots.tip ? slots.tip() : props.tip && <div class={`${CLASS_PREFIX}__tip`}>{props.tip}</div>}
+        {slots.tip ? slots.tip() : props.tip && <div class={`${resolveClassName(CLASS_PREFIX)}__tip`}>{props.tip}</div>}
         <UploadList
           files={fileList.value}
           theme={props.theme}