+++ site/changelog.md
@@ -1,5 +1,12 @@
 ## 更新日志
 
+### 2.0.1-beta.6
+
+* **[feat]**:
+    - feat(table): 增加 checked prop配置说明 & is-selected-fn 完善参数 & 增加props.checked说明
+    - 修复 Cascader、ColorPicker、Transfer、Upload、Tree 组件部分样式无 prefix 设置的问题
+
+
 ### 2.0.1-beta.5
 
 * **[feat]**: